backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
docarray
pypdf
python-dotenv
tiktoken
